feat : update solution & SimpleStore proj 1) Add /tests folder to solution. 2) Refactoring of SimpleStore proj. 3) Update tests for SimpleStore proj.
