+++ Aslamov.TMS0212.sln
@@ -1,0 +1,47 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleStoreApp", "src\SimpleStore\SimpleStoreApp\SimpleStoreApp.csproj", "{1426BC3B-4127-4C6F-B920-C7FD5E80590F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleStore", "src\SimpleStore\SimpleStore\SimpleStore.csproj", "{47F0F5A3-EE6D-4E85-BDE6-05B43E8EFB48}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleStoreTests", "tests\SimpleStoreTests\SimpleStoreTests.csproj", "{49D6F41C-8803-4E8A-8931-2B3CE908BEB7}"
+EndProject
+Project("{9344BDBB-3E7F-41FC-A0DD-8665D75EE146}") = "HelloWorld", "src\HelloWorld\HelloWorld\HelloWorld.csproj", "{94FC90C6-9E2D-4090-A66F-32C67A836D94}"
+EndProject
+Project("{9344BDBB-3E7F-41FC-A0DD-8665D75EE146}") = "SimpleCalc", "src\SimpleCalc\SimpleCalc.csproj", "{3A0EB77F-1297-4810-8715-7D68A5A380F3}"
+EndProject
+Project("{9344BDBB-3E7F-41FC-A0DD-8665D75EE146}") = "SimpleCalendar", "src\SimpleCalendar\SimpleCalendar.csproj", "{8BE32B1E-DB48-43D8-8426-E76FC6751CCF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|anycpu = Debug|anycpu
+		Release|anycpu = Release|anycpu
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1426BC3B-4127-4C6F-B920-C7FD5E80590F}.Debug|anycpu.ActiveCfg = Debug|Any CPU
+		{1426BC3B-4127-4C6F-B920-C7FD5E80590F}.Debug|anycpu.Build.0 = Debug|Any CPU
+		{1426BC3B-4127-4C6F-B920-C7FD5E80590F}.Release|anycpu.ActiveCfg = Release|Any CPU
+		{1426BC3B-4127-4C6F-B920-C7FD5E80590F}.Release|anycpu.Build.0 = Release|Any CPU
+		{47F0F5A3-EE6D-4E85-BDE6-05B43E8EFB48}.Debug|anycpu.ActiveCfg = Debug|Any CPU
+		{47F0F5A3-EE6D-4E85-BDE6-05B43E8EFB48}.Debug|anycpu.Build.0 = Debug|Any CPU
+		{47F0F5A3-EE6D-4E85-BDE6-05B43E8EFB48}.Release|anycpu.ActiveCfg = Release|Any CPU
+		{47F0F5A3-EE6D-4E85-BDE6-05B43E8EFB48}.Release|anycpu.Build.0 = Release|Any CPU
+		{49D6F41C-8803-4E8A-8931-2B3CE908BEB7}.Debug|anycpu.ActiveCfg = Debug|Any CPU
+		{49D6F41C-8803-4E8A-8931-2B3CE908BEB7}.Debug|anycpu.Build.0 = Debug|Any CPU
+		{49D6F41C-8803-4E8A-8931-2B3CE908BEB7}.Release|anycpu.ActiveCfg = Release|Any CPU
+		{49D6F41C-8803-4E8A-8931-2B3CE908BEB7}.Release|anycpu.Build.0 = Release|Any CPU
+		{94FC90C6-9E2D-4090-A66F-32C67A836D94}.Debug|anycpu.ActiveCfg = Debug|anycpu
+		{94FC90C6-9E2D-4090-A66F-32C67A836D94}.Debug|anycpu.Build.0 = Debug|anycpu
+		{94FC90C6-9E2D-4090-A66F-32C67A836D94}.Release|anycpu.ActiveCfg = Release|anycpu
+		{94FC90C6-9E2D-4090-A66F-32C67A836D94}.Release|anycpu.Build.0 = Release|anycpu
+		{3A0EB77F-1297-4810-8715-7D68A5A380F3}.Debug|anycpu.ActiveCfg = Debug|Any CPU
+		{3A0EB77F-1297-4810-8715-7D68A5A380F3}.Debug|anycpu.Build.0 = Debug|Any CPU
+		{3A0EB77F-1297-4810-8715-7D68A5A380F3}.Release|anycpu.ActiveCfg = Release|Any CPU
+		{3A0EB77F-1297-4810-8715-7D68A5A380F3}.Release|anycpu.Build.0 = Release|Any CPU
+		{8BE32B1E-DB48-43D8-8426-E76FC6751CCF}.Debug|anycpu.ActiveCfg = Debug|Any CPU
+		{8BE32B1E-DB48-43D8-8426-E76FC6751CCF}.Debug|anycpu.Build.0 = Debug|Any CPU
+		{8BE32B1E-DB48-43D8-8426-E76FC6751CCF}.Release|anycpu.ActiveCfg = Release|Any CPU
+		{8BE32B1E-DB48-43D8-8426-E76FC6751CCF}.Release|anycpu.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
