--- src/HelloWorld/HelloWorld.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloWorld", "HelloWorld\HelloWorld.csproj", "{D3052818-BBC9-4001-97EF-8324070FC8B8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|anycpu = Debug|anycpu
-		Release|anycpu = Release|anycpu
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D3052818-BBC9-4001-97EF-8324070FC8B8}.Debug|anycpu.ActiveCfg = Debug|anycpu
-		{D3052818-BBC9-4001-97EF-8324070FC8B8}.Debug|anycpu.Build.0 = Debug|anycpu
-		{D3052818-BBC9-4001-97EF-8324070FC8B8}.Release|anycpu.ActiveCfg = Release|anycpu
-		{D3052818-BBC9-4001-97EF-8324070FC8B8}.Release|anycpu.Build.0 = Release|anycpu
-	EndGlobalSection
-EndGlobal
