--- src/SimpleCalc/SimpleCalc.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleCalc", "SimpleCalc.csproj", "{FE5F5C24-C054-47BD-852C-3276A6F23D29}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FE5F5C24-C054-47BD-852C-3276A6F23D29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FE5F5C24-C054-47BD-852C-3276A6F23D29}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FE5F5C24-C054-47BD-852C-3276A6F23D29}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FE5F5C24-C054-47BD-852C-3276A6F23D29}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
