--- src/SimpleCalendar/SimpleCalendar.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleCalendar", "SimpleCalendar.csproj", "{6182E2FB-616C-4B71-8991-5E3EF8567D17}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6182E2FB-616C-4B71-8991-5E3EF8567D17}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6182E2FB-616C-4B71-8991-5E3EF8567D17}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6182E2FB-616C-4B71-8991-5E3EF8567D17}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6182E2FB-616C-4B71-8991-5E3EF8567D17}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
