--- src/SimpleStore/SimpleStore.sln
@@ -1,23 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleStore", "SimpleStore.csproj", "{5509BB2E-C294-41ED-8FDD-B5E56AE660D4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleStoreTests", "SimpleStoreTests\SimpleStoreTests.csproj", "{B0178BA2-9D44-49CF-8A59-7D36F5E19839}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5509BB2E-C294-41ED-8FDD-B5E56AE660D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5509BB2E-C294-41ED-8FDD-B5E56AE660D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5509BB2E-C294-41ED-8FDD-B5E56AE660D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5509BB2E-C294-41ED-8FDD-B5E56AE660D4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0178BA2-9D44-49CF-8A59-7D36F5E19839}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0178BA2-9D44-49CF-8A59-7D36F5E19839}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0178BA2-9D44-49CF-8A59-7D36F5E19839}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0178BA2-9D44-49CF-8A59-7D36F5E19839}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
