+++ src/SimpleStore/SimpleStore/SimpleStore.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Store\" />
+    <Folder Include="Product\" />
+  </ItemGroup>
+</Project>
