+++ tests/SimpleStoreTests/SimpleStoreTests.csproj
@@ -14,6 +14,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SimpleStore.csproj" />
+    <ProjectReference Include="..\..\src\SimpleStore\SimpleStore\SimpleStore.csproj" />
   </ItemGroup>
 </Project>
